Adding lifeline and slowing removing pencil out of view
