+++ PencilAdventure/PencilAdventure/Images.xcassets/house_background.imageset/Contents.json
@@ -13,6 +13,11 @@
       "filename" : "house_background@2x~iphone.png"
     },
     {
+      "idiom" : "iphone",
+      "subtype" : "retina4",
+      "scale" : "2x"
+    },
+    {
       "idiom" : "ipad"
     },
     {
@@ -28,6 +33,6 @@
   ],
   "info" : {
     "version" : 1,
-    "author" : "artgen"
+    "author" : "xcode"
   }
 }
